refactor: notification oragnization
