+++ packages/notification/index.ts
@@ -1,0 +1,11 @@
+import Notification from './notification'
+import NotificationWithLanguage from './notification-with-language'
+import NotificationWithProduct from './notification-with-product'
+import NotificationWithResource from './notification-with-resource'
+
+export default Notification
+export {
+  NotificationWithLanguage,
+  NotificationWithProduct,
+  NotificationWithResource,
+}
