+++ packages/notification/types.ts
@@ -1,3 +1,4 @@
+import * as React from 'react'
 import type { Products } from '@hashicorp/platform-product-meta'
 
 export type NotificationLanguages = 'en' | 'de' | 'fr' | 'jp' | 'kr'
@@ -6,17 +7,26 @@ export type NotificationTypes = 'podcast' | 'webinar' | 'whitepaper'
 
 export interface NotificationProps {
   appearance?: 'light' | 'dark'
-  language?: NotificationLanguages
-  type?: NotificationTypes
-  thumbnail?: {
-    src: string
-    alt: string
-  }
-  product?: NotificationProducts
   description: string
   onDismiss: () => void
   cta: {
     title: string
     url: string
   }
+  children?: React.ReactNode
+}
+
+export interface NotificationWithLanguageProps
+  extends Exclude<NotificationProps, 'children'> {
+  language: NotificationLanguages
+}
+
+export interface NotificationWithProductProps
+  extends Exclude<NotificationProps, 'children'> {
+  product: NotificationProducts
+}
+
+export interface NotificationWithResourceProps
+  extends Exclude<NotificationProps, 'children'> {
+  type: NotificationTypes
 }
